+++ redpandas/redpd_df.py
@@ -107,28 +107,104 @@ def redpd_dataframe(input_dw: DataWindow,
 
     return df_all_sensors_all_stations
 
+# def three_step_to_flat(df: pd.DataFrame,
+#                        name_column: str):
+#     """
+#
+#     :param df: input pandas DataFrame. REQUIRED
+#     :param name_column: string, name of column
+#     :return: nothing, acts on the df
+#     """
+#     # Create new columns with shape tuple for future unflattening/reshaping
+#     df[[f'{name_column}_ndim']] = df[[f'{name_column}']].applymap(np.shape)
+#     # Change tuples to 1D np.array to save it to parquet
+#     df[[f'{name_column}_ndim']] = df[[f'{name_column}_ndim']].applymap(np.asarray)
+#     # Flatten each row in wf columns
+#     df[[f'{name_column}']] = df[[f'{name_column}']].applymap(np.ravel)
 
-def three_step_to_flat(df: pd.DataFrame,
-                       name_column: str):
-    """
 
-    :param df: input pandas DataFrame. REQUIRED
-    :param name_column: string, name of column
-    :return: nothing, acts on the df
-    """
-    # Create new columns with shape tuple for future unflattening/reshaping
-    df[[f'{name_column}_ndim']] = df[[f'{name_column}']].applymap(np.shape)
-    # Change tuples to 1D np.array to save it to parquet
-    df[[f'{name_column}_ndim']] = df[[f'{name_column}_ndim']].applymap(np.asarray)
-    # Flatten each row in wf columns
-    df[[f'{name_column}']] = df[[f'{name_column}']].applymap(np.ravel)
+# def export_df_to_parquet(df: pd.DataFrame,
+#                          output_dir_pqt: str,
+#                          tfr_column_label: Union[List[str], str, None] = None,
+#                          tfr_frequency_label: Union[List[str], str, None] = None,
+#                          tfr_time_label: Union[List[str], str, None] = None,
+#                          output_filename_pqt: Optional[str] = None,
+#                          event_name: Optional[str] = "Redvox") -> str:
+#     """
+#     Export RedPandas DataFrame to parquet
+#
+#     :param df: input pandas DataFrame. REQUIRED
+#     :param output_dir_pqt: string, output directory for parquet. REQUIRED
+#     :param tfr_column_label: string or list of strings, column label with TFR
+#     :param tfr_frequency_label: string or list of strings, column label with TFR frequencies
+#     :param tfr_time_label: string or list of strings, column label with TFR time
+#     :param output_filename_pqt: optional string for parquet filename. Default is None
+#     :param event_name: optional string with name of event. Default is "Redvox"
+#
+#     :return: string with full path (output directory and filename) of parquet
+#     """
+#     if not os.path.exists(output_dir_pqt):  # make output directory if it doesn't exist
+#         print(f"Creating output directory: {output_dir_pqt}...")
+#         os.mkdir(output_dir_pqt)
+#
+#     key_sensors = {'barometer', 'accelerometer', 'gyroscope', 'magnetometer'}
+#     for label in key_sensors:
+#         # Create new columns with shape tuple for future unflattening/reshaping
+#         if f'{label}_wf_raw' in df.columns:
+#             three_step_to_flat(df=df,
+#                                name_column=f'{label}_wf_raw')
+#
+#         if f'{label}_wf_highpass' in df.columns:
+#             three_step_to_flat(df=df,
+#                                name_column=f'{label}_wf_highpass')
+#
+#         if f'{label}_nans' in df.columns:
+#             three_step_to_flat(df=df,
+#                                name_column=f'{label}_nans')
+#
+#     if not isinstance(tfr_column_label, list):
+#         tfr_column_label = [tfr_column_label]
+#     if not isinstance(tfr_frequency_label, list):
+#         tfr_frequency_label = [tfr_frequency_label]
+#     if not isinstance(tfr_time_label, list):
+#         tfr_time_label = [tfr_time_label]
+#
+#     for index_list, tfr_columns_in_df in enumerate(tfr_column_label):
+#         tfr_frequency_in_df = tfr_frequency_label[index_list]
+#         tfr_time_in_df = tfr_time_label[index_list]
+#
+#         if tfr_columns_in_df is not None:
+#             if tfr_columns_in_df in df.columns:
+#                 three_step_to_flat(df=df,
+#                                    name_column=f'{tfr_columns_in_df}')
+#
+#         if tfr_frequency_in_df is not None:
+#             if tfr_frequency_in_df in df.columns:
+#                 three_step_to_flat(df=df,
+#                                    name_column=f'{tfr_frequency_in_df}')
+#
+#         if tfr_time_in_df is not None:
+#             if tfr_time_in_df in df.columns:
+#                 three_step_to_flat(df=df,
+#                                    name_column=f'{tfr_time_in_df}')
+#
+#     # Make filename if non given
+#     if output_filename_pqt is None:
+#         output_filename_pqt: str = event_name + "_df.parquet"
+#
+#     if output_filename_pqt.find(".parquet") == -1 and output_filename_pqt.find(".pqt") == -1:
+#         full_output_dir_path_parquet = os.path.join(output_dir_pqt, output_filename_pqt + ".parquet")
+#     else:
+#         full_output_dir_path_parquet = os.path.join(output_dir_pqt, output_filename_pqt)
+#
+#     df.to_parquet(full_output_dir_path_parquet)
+#     print(f"\nExported Parquet RedPandas DataFrame to {full_output_dir_path_parquet}")
+#
+#     return full_output_dir_path_parquet
 
 
 def export_df_to_parquet(df: pd.DataFrame,
                          output_dir_pqt: str,
-                         tfr_column_label: Union[List[str], str, None] = None,
-                         tfr_frequency_label: Union[List[str], str, None] = None,
-                         tfr_time_label: Union[List[str], str, None] = None,
                          output_filename_pqt: Optional[str] = None,
                          event_name: Optional[str] = "Redvox") -> str:
     """
@@ -136,58 +212,24 @@ def export_df_to_parquet(df: pd.DataFrame,
 
     :param df: input pandas DataFrame. REQUIRED
     :param output_dir_pqt: string, output directory for parquet. REQUIRED
-    :param tfr_column_label: string or list of strings, column label where TFR
-    :param tfr_frequency_label:
-    :param tfr_time_label:
     :param output_filename_pqt: optional string for parquet filename. Default is None
     :param event_name: optional string with name of event. Default is "Redvox"
 
     :return: string with full path (output directory and filename) of parquet
     """
-    if not os.path.exists(output_dir_pqt):  # make output directory if it doesn't exist
-        print(f"Creating output directory: {output_dir_pqt}...")
-        os.mkdir(output_dir_pqt)
 
-    key_sensors = {'barometer', 'accelerometer', 'gyroscope', 'magnetometer'}
-    for label in key_sensors:
-        # Create new columns with shape tuple for future unflattening/reshaping
-        if f'{label}_wf_raw' in df.columns:
-            three_step_to_flat(df=df,
-                               name_column=f'{label}_wf_raw')
+    for column in df.columns:
+        for row in df.index:  # check it is array by cheking all rows, look into dtypes
+            check = np.shape(df[column][row])
+            if len(check) >= 2:
+                # Create new columns with shape tuple for future unflattening/reshaping
+                df[[f'{column}_ndim']] = df[[f'{column}']].applymap(np.shape)
+                # Change tuples to 1D np.array to save it to parquet
+                df[[f'{column}_ndim']] = df[[f'{column}_ndim']].applymap(np.asarray)
+                # Flatten each row in wf columns
+                df[[f'{column}']] = df[[f'{column}']].applymap(np.ravel)
 
-        if f'{label}_wf_highpass' in df.columns:
-            three_step_to_flat(df=df,
-                               name_column=f'{label}_wf_highpass')
-
-        if f'{label}_nans' in df.columns:
-            three_step_to_flat(df=df,
-                               name_column=f'{label}_nans')
-
-    if isinstance(tfr_column_label, str):
-        tfr_column_label = [tfr_column_label]
-    if isinstance(tfr_frequency_label, str):
-        tfr_frequency_label = [tfr_frequency_label]
-    if isinstance(tfr_time_label, str):
-        tfr_time_label = [tfr_time_label]
-
-    for index_list, tfr_columns_in_df in enumerate(tfr_column_label):
-        tfr_frequency_in_df = tfr_frequency_label[index_list]
-        tfr_time_in_df = tfr_time_label[index_list]
-
-        if tfr_columns_in_df is not None:
-            if tfr_columns_in_df in df.columns:
-                three_step_to_flat(df=df,
-                                   name_column=f'{tfr_columns_in_df}')
-
-        if tfr_frequency_in_df is not None:
-            if tfr_frequency_in_df in df.columns:
-                three_step_to_flat(df=df,
-                                   name_column=f'{tfr_frequency_in_df}')
-
-        if tfr_time_in_df is not None:
-            if tfr_time_in_df in df.columns:
-                three_step_to_flat(df=df,
-                                   name_column=f'{tfr_time_in_df}')
+                break
 
     # Make filename if non given
     if output_filename_pqt is None:
@@ -202,3 +244,4 @@ def export_df_to_parquet(df: pd.DataFrame,
     print(f"\nExported Parquet RedPandas DataFrame to {full_output_dir_path_parquet}")
 
     return full_output_dir_path_parquet
+