+++ redpandas/tests/boom.py
@@ -6,7 +6,7 @@ from redpandas.redpd_plot.wiggles import plot_wiggles_pandas
 import redpandas.redpd_tfr as rpd_tfr
 from redpandas.redpd_plot.mesh import plot_mesh_pandas
 import redpandas.redpd_dq as rpd_dq
-from redpandas.redpd_preprocess import df_column_unflatten
+from redpandas.redpd_preprocess import df_column_unflatten, df_unflatten
 
 # from redvox.common.data_window import DataWindow
 from redvox.common.data_window import DataWindow, EventOrigin
@@ -29,52 +29,43 @@ if __name__ == '__main__':
 
     DWAConfig = DataWindowConfig(input_dir=INPUT_DIR,
                                  station_ids=[
-                                     "1637610015",
-                                              "2551278155",
-                                              "1637610012",
-                                              "1637610015",
-                                              "1637610014",
-                                              "872266036"
+                                     "1637610001",
+                                      "2551278155",
+                                      "1637610012",
+                                      "1637610015",
+                                      "1637610014",
+                                      "872266036"
                                               ],
                                  start_datetime=dt.datetime_from_epoch_seconds_utc(1632006000),
                                  end_datetime=dt.datetime_from_epoch_seconds_utc(1632006330))
 
     rdvx_data: DataWindow = DataWindow(event_name="dw",
-                                       # event_origin=DWOrigin,
                                        config=DWAConfig,
                                        out_dir="/Users/meritxell/Desktop/test",
                                        out_type="lz4",
                                        debug=True)
-    # path = rdvx_dataa.save()
-    # print(path)
-    # station = rdvx_data.get_station("0872266036")[0]
-    # print(station.health_sensor().get_data_channel("network_type"))
-    #
-    # exit()
 
     df0 = redpd_dataframe(input_dw=rdvx_data,
-                              sensor_labels=["audio",
-                                             "barometer",
-                                             "accelerometer",
-                                             "gyroscope",
-                                             "magnetometer",
-                                             "clock",
-                                             "synchronization",
-                                             "location",
-                                             "health"
-                                             ])
+                          sensor_labels=["audio",
+                                         "barometer",
+                                         "accelerometer",
+                                         "gyroscope",
+                                         "magnetometer",
+                                         "clock",
+                                         "synchronization",
+                                         "location",
+                                         "health"
+                                         ])
 
-
-    # # Check audio wiggles ok
-    # fig_wiggles = plot_wiggles_pandas(df=df0,
-    #                                   sig_wf_label=["audio_wf"],
-    #                                   sig_timestamps_label=["audio_epoch_s"],
-    #                                   sig_id_label="station_id",
-    #                                   station_id_str="1637610078",
-    #                                   show_figure=False)
+    # Check audio wiggles ok
+    fig_wiggles = plot_wiggles_pandas(df=df0,
+                                      sig_wf_label=["audio_wf"],
+                                      sig_timestamps_label=["audio_epoch_s"],
+                                      sig_id_label="station_id",
+                                      show_figure=False)
 
     # TFR
-    sensors_to_tfr = ["audio", 'accelerometer'] #"barometer", "accelerometer", "magnetometer", "gyroscope"]
+    sensors_to_tfr = ["audio", 'accelerometer']  # "barometer", "accelerometer", "magnetometer", "gyroscope"]
     for sensor in sensors_to_tfr:
         if sensor == "audio":
             sensor_wf = "audio_wf"
@@ -92,28 +83,22 @@ if __name__ == '__main__':
                                      new_column_tfr_frequency_hz=f"{sensor}_tfr_frequency_hz",
                                      new_column_tfr_time_s=f"{sensor}_tfr_time_s")
 
-
     export_df_to_parquet(df=df0,
-                         output_dir_pqt="/Users/meritxell/Desktop",
-                         tfr_column_label=["audio_tfr_bits", "accelerometer_tfr_bits"],
-                         tfr_frequency_label=["audio_tfr_frequency_hz", "accelerometer_tfr_frequency_hz"],
-                         tfr_time_label=["audio_tfr_time_s", "accelerometer_tfr_time_s"])
+                         output_dir_pqt="/Users/meritxell/Desktop")
 
-    # df_sensors = pd.read_parquet("/Users/meritxell/Desktop/Redvox_df.parquet")
-    #
+    df_sensors = pd.read_parquet("/Users/meritxell/Desktop/Redvox_df.parquet")
+
     # df_column_unflatten(df=df_sensors,
     #                     col_wf_label="accelerometer_tfr_bits",
     #                     col_ndim_label="accelerometer_tfr_bits_ndim")
 
-    exit()
-    print(rdvx_data.event_name)
-    print(rdvx_data)
+    df_unflatten(df_sensors)
 
-    rpd_dq.mic_sync(rdvx_data)
-    rpd_dq.station_channel_timing(rdvx_data)
-    rpd_dq.station_metadata(rdvx_data)
+    # rpd_dq.mic_sync(rdvx_data)
+    # rpd_dq.station_channel_timing(rdvx_data)
+    # rpd_dq.station_metadata(rdvx_data)
 
-    fig_mesh = plot_mesh_pandas(df=df0,
+    fig_mesh = plot_mesh_pandas(df=df_sensors,
                                 mesh_time_label="audio_tfr_time_s",
                                 mesh_frequency_label="audio_tfr_frequency_hz",
                                 mesh_tfr_label="audio_tfr_bits",
@@ -128,15 +113,15 @@ if __name__ == '__main__':
                                 mesh_color_range=[16.0, 16.0, 16.0, 16.0, 16.0],
                                 show_figure=True,
                                 ytick_values_show=True)
-    for station in df0.index:
-        if df0["station_id"][station] == "0872266036":
-            pnl.plot_wf_mesh_vert(redvox_id=df0["station_id"][station],
-                                  wf_panel_2_sig=df0["audio_wf"][station],
-                                  wf_panel_2_time=df0["audio_epoch_s"][station],
-                                  mesh_time=df0["audio_tfr_time_s"][station],
-                                  mesh_frequency=df0["audio_tfr_frequency_hz"][station],
-                                  mesh_panel_0_tfr=df0["audio_tfr_bits"][station],
-                                  start_time_epoch=df0["audio_epoch_s"][0][0],
+    for station in df_sensors.index:
+        if df_sensors["station_id"][station] == "0872266036":
+            pnl.plot_wf_mesh_vert(redvox_id=df_sensors["station_id"][station],
+                                  wf_panel_2_sig=df_sensors["audio_wf"][station],
+                                  wf_panel_2_time=df_sensors["audio_epoch_s"][station],
+                                  mesh_time=df_sensors["audio_tfr_time_s"][station],
+                                  mesh_frequency=df_sensors["audio_tfr_frequency_hz"][station],
+                                  mesh_panel_0_tfr=df_sensors["audio_tfr_bits"][station],
+                                  start_time_epoch=df_sensors["audio_epoch_s"][0][0],
                                   frequency_scaling="log",
                                   frequency_hz_ymax=320,
                                   frequency_hz_ymin=20,
@@ -146,13 +131,13 @@ if __name__ == '__main__':
                                   figure_title=f"STFT")
 
         else:
-            pnl.plot_wf_mesh_vert(redvox_id=df0["station_id"][station],
-                                  wf_panel_2_sig=df0["audio_wf"][station],
-                                  wf_panel_2_time=df0["audio_epoch_s"][station],
-                                  mesh_time=df0["audio_tfr_time_s"][station],
-                                  mesh_frequency=df0["audio_tfr_frequency_hz"][station],
-                                  mesh_panel_0_tfr=df0["audio_tfr_bits"][station],
-                                  start_time_epoch=df0["audio_epoch_s"][0][0],
+            pnl.plot_wf_mesh_vert(redvox_id=df_sensors["station_id"][station],
+                                  wf_panel_2_sig=df_sensors["audio_wf"][station],
+                                  wf_panel_2_time=df_sensors["audio_epoch_s"][station],
+                                  mesh_time=df_sensors["audio_tfr_time_s"][station],
+                                  mesh_frequency=df_sensors["audio_tfr_frequency_hz"][station],
+                                  mesh_panel_0_tfr=df_sensors["audio_tfr_bits"][station],
+                                  start_time_epoch=df_sensors["audio_epoch_s"][0][0],
                                   frequency_scaling="log",
                                   mesh_panel_0_colormap_scaling="range",
                                   mesh_panel_0_color_range=16.,
@@ -160,30 +145,15 @@ if __name__ == '__main__':
                                   figure_title=f"STFT")
 
     # Check barometer wiggles
-    fig_wiggles_bar = plot_wiggles_pandas(df=df0,
+    fig_wiggles_bar = plot_wiggles_pandas(df=df_sensors,
                                           sig_wf_label=["barometer_wf_highpass"],
                                           sig_timestamps_label=["barometer_epoch_s"],
                                           sig_id_label="station_id",
-                                          show_figure=False)
+                                          show_figure=True)
     # Check accelerometer wiggles
-    fig_wiggles_acc = plot_wiggles_pandas(df=df0,
+    fig_wiggles_acc = plot_wiggles_pandas(df=df_sensors,
                                           sig_wf_label=["accelerometer_wf_highpass"],
                                           sig_timestamps_label=["accelerometer_epoch_s"],
                                           sig_id_label="station_id",
-                                          show_figure=False)
-
-    for station in df0.index:
-        if type(df0["accelerometer_wf_highpass"][station]) != float:
-            pnl.plot_wf_wf_wf_vert(redvox_id=df0["station_id"][station],
-                                   wf_panel_2_sig=df0["audio_wf"][station],
-                                   wf_panel_2_time=df0["audio_epoch_s"][station],
-                                   wf_panel_1_sig=df0["barometer_wf_highpass"][station][0],
-                                   wf_panel_1_time=df0["barometer_epoch_s"][station],
-                                   wf_panel_0_sig=df0["accelerometer_wf_highpass"][station][2],
-                                   wf_panel_0_time=df0["accelerometer_epoch_s"][station],
-                                   start_time_epoch=df0["audio_epoch_s"][station][0],
-                                   wf_panel_2_units="Mic, Norm",
-                                   wf_panel_1_units="Bar hp, kPa",
-                                   wf_panel_0_units="Acc Z hp, m/$s^2$")
-
+                                          show_figure=True)
     plt.show()